Fix stale document fragments by md5-hashing content
Previously, we stored document fragments with IDs like
<title>:<position-in-document>. This provided no way to resolve stale
data if a fragment moved from one page to another or disappeared
completely from an indexed site.

Now, we index pages and sections with an ID derived from an md5 hash of
text of a fragment and page's title and content.

We use these md5 hashes to resolve stale data after indexing and delete
any outdated Hashes.