Include the page title in the section ID hash
Currently, the section ID hash does not include the page title.
This means that if the page title changes, section documents will
get left behind, becoming stale.

By hashing the page title with the other fields we hash for section
IDs, changing a page's title will expunge section documents. This is
important because section documents include the page title -- so if
it changes, we need to rebuild these section documents, so that they
include the new page title.